Move things around and create textures module
