--- src/rect.spec.js
@@ -1,70 +1,0 @@
-const Rect = require("./Rect");
-
-describe("Rect", () => {
-  test("It should have a Rect function", () => {
-    expect(typeof Rect).toBe("function");
-  });
-
-  test("The Rect function should build a rect js object", () => {
-    const expected = {
-      type: "RECTANGLE",
-      position: { x: 0, y: 0, z: 0, width: 5 },
-      children: []
-    };
-
-    const actual = Rect({ x: 0, y: 0, z: 0, width: 5, children: [] });
-
-    expect(actual).toEqual(expected);
-  });
-
-  test("The Rect function should take children", () => {
-    const expected = {
-      type: "RECTANGLE",
-      position: { x: 0, y: 0, z: 0, width: 5 },
-      children: [
-        {
-          type: "RECTANGLE",
-          position: { x: 0, y: 0, z: 0, width: 5 },
-          children: []
-        }
-      ]
-    };
-
-    const actual = Rect({
-      x: 0,
-      y: 0,
-      z: 0,
-      width: 5,
-      children: [Rect({ x: 0, y: 0, z: 0, width: 5, children: [] })]
-    });
-
-    expect(actual).toEqual(expected);
-  });
-
-  test("The node functions should be able to render custom logic nodes", () => {
-    const expected = {
-      type: "RECTANGLE",
-      position: { x: 1, y: 1, z: 1, width: 5 },
-      children: []
-    };
-
-    const customElement = props =>
-      Rect({
-        x: props.x / 2,
-        y: props.y / 2,
-        z: props.z / 2,
-        width: props.width / 2,
-        children: []
-      });
-
-    const actual = customElement({
-      x: 2,
-      y: 2,
-      z: 2,
-      width: 10,
-      children: []
-    });
-
-    expect(actual).toEqual(expected);
-  });
-});
