+++ src/renderer/__image_snapshots__/renderer-spec-js-renderer-renderer-should-apply-translation-to-grand-children-1-snap.png
