+++ src/renderer/__image_snapshots__/renderer-spec-js-renderer-renderer-should-render-a-rect-1-snap.png
