+++ src/renderer/__image_snapshots__/renderer-spec-js-renderer-renderer-should-render-a-rect-with-rotation-1-snap.png
