+++ src/renderer/__image_snapshots__/renderer-spec-js-renderer-renderer-should-render-static-sprites-1-snap.png
